requirements.txt
aiogram==2.25.1
SQLAlchemy==1.4.44
psycopg2-binary==2.9.5
python-dotenv==0.21.0
